Add policy to resources

+++ app/Http/Controllers/ResourceTypeController.php
@@ -17,6 +17,11 @@ class ResourceTypeController extends Controller
 
     public function index()
     {
+        if (\Gate::denies('read', 'resources')) {
+            flash("You don't have access to this page");
+            return \Redirect::to('/');
+        }
+
         $resource_levels = ResourceType::tree()->orderBy('name')->paginate();
         $resources = Resources::lists('id', 'name')->all();
 
@@ -25,6 +30,11 @@ class ResourceTypeController extends Controller
 
     public function create()
     {
+        if (\Gate::denies('write', 'resources')) {
+            flash("You don't have access to this page");
+            return \Redirect::to('/');
+        }
+
         $resource_types = ResourceType::options();
         return view('resource-type.create', compact('resource_types'));
     }
@@ -34,6 +44,10 @@ class ResourceTypeController extends Controller
      */
     public function store(Request $request)
     {
+        if (\Gate::denies('write', 'resources')) {
+            flash("You don't have access to this page");
+            return \Redirect::to('/');
+        }
 
         $this->validate($request, $this->rules);
         ResourceType::create([
@@ -47,6 +61,10 @@ class ResourceTypeController extends Controller
 
     public function show(ResourceType $resource_type)
     {
+        if (\Gate::denies('read', 'resources')) {
+            flash("You don't have access to this page");
+            return \Redirect::to('/');
+        }
 
         $resource_types = ResourceType::options();
         return view('resource-type.show', compact('resource_type','resource_types'));
@@ -54,12 +72,22 @@ class ResourceTypeController extends Controller
 
     public function edit(ResourceType $resource_type)
     {
+        if (\Gate::denies('write', 'resources')) {
+            flash("You don't have access to this page");
+            return \Redirect::to('/');
+        }
+
         $resource_types = ResourceType::options();
         return view('resource-type.edit', compact('resource_types', 'resource_type'));
     }
 
     public function update(ResourceType $resource_type, Request $request)
     {
+        if (\Gate::denies('write', 'resources')) {
+            flash("You don't have access to this page");
+            return \Redirect::to('/');
+        }
+
         $this->validate($request, $this->rules);
         $resource_type->update($request->all());
         flash('Resource type has been saved', 'success');
@@ -68,10 +96,16 @@ class ResourceTypeController extends Controller
 
     public function destroy(ResourceType $resource_type)
     {
+        if (\Gate::denies('delete', 'resources')) {
+            flash("You don't have access to this page");
+            return \Redirect::to('/');
+        }
+
         $resource_type->delete();
         flash('Resource type has been deleted', 'success');
         return \Redirect::route('resource-type.index');
     }
+
     function wipe(WipeRequest $request)
     {
         \DB::table('resource_types')->delete();
